First-commit

+++ src/main/java/com/gizasystems/filemanagement/infrastructure/FileSystemConfig.java
@@ -17,6 +17,6 @@ public class FileSystemConfig {
 
 
 
-    private long maxFileSize;
+    private int maxFileSize;
     private String storagePath;
 }
